fixed VBR stuffing bits problem. if VBR uses a large frame, then ResvMax may become small, and we have to make sure that main_data_begin doesn't create a buffer which violates this condition.
The old code wouldn't check the above, but it would pad the current
frame with 0's (wasting bits) to make sure the *next* frame didn't
violate the ResvMax condition.  But this results in wasted bits
because in VBR there is a change the *next* frame will be at
a lower bitrate and thus have a larger ResvMax.


git-svn-id: https://svn.code.sf.net/p/lame/svn/trunk@529 981c7233-bbb0-e711-95a9-001517a2e1a4