possible serious buffering problem fixed.  LAME buffers input data until it has at least 1904 samples.  It then starts processing them.
The old MDCT routines processed 1728 samples and had a delay of 528.
(introduced in 3.54) But the new MDCT routines, because the delay is
so small (48 samples) actually need 2014 samples to be in the buffer!

If you use the lame input routines, for MPEG1, this bug would
not be found because the buffer always has well over 2014 samples
in it.  But it was possible to trigger it in MPEG2 encodings.
It would also be possible to trigger this bug in programs
that use lame_enc.dll and pass LAME data that is not in
chunks of 1152 samples.


git-svn-id: https://svn.code.sf.net/p/lame/svn/trunk@627 981c7233-bbb0-e711-95a9-001517a2e1a4