started adding more prototypes into lame.h:
int lame_set_variable_name(lame_global_flags *, int)
int lame_get_variable_name(lame_global_flags *)

_get will return the value (instead of an error code)
as suggested.

_set will return an error code.  But right now, the way lame
works there is no error checking until lame_init_params() is
called.  This is because, for example, bitrate=135
is not an error if free_format=1, but it is an error
for CBR.  So many errors cannot be detected until all options
are set.