spreading function moved into seperate routine, s3_func(). s3 normalized so that:
         +inf
           /
           |  s3 [ bark ]  d(bark)   =  1
           /
         -inf

which resulted in dividing the result by 0.66.

This is the first step in fixing the long outstanding bug
mentioned by Takehiro:  s3[i][j] is normalized improperly for
values of i near the partition band boundaries.  For these values,
the computed 'norm' is much too small since a lot of the data is
missing, resulting in artificially increasing the spreading function
(and thus the masking)

--nspsytune: s3_l function is used with a normalization of .5
(instead of computing the norm).  So I changed this to a .5*.66
to compensate for the above change.

Next step:

1.
s3[i][j] will be multiplied by the width (in barks) of the
j'th partition band (about .35).  Once this is done, we will not need to
compute and divide by 'norm' later.


2.
Make NOTABLES the default

3.
compute psymodel all the way up to 22khz.