added lame_encode_flush_nogap().
This routine will flush all internal mp3 buffers, and pad the
last frame with ancillary data so it is a complete mp3 frame.

It will then set ResvSize to 0, so we can continue encoding.

after a call to lame_encode_flush_nogap(), if we close
the mp3 file and open a new one for the future data,
then the two mp3 files should play back 'gapless'
if they are first concatenated together.


git-svn-id: https://svn.code.sf.net/p/lame/svn/trunk@2103 981c7233-bbb0-e711-95a9-001517a2e1a4